requirements.txt
asgiref==3.3.0
astroid==2.4.2
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.0.10
django-admin-sortable2==0.7.7
django-allauth==0.43.0
django-angular==2.3
django-appconf==1.0.4
django-classy-tags==2.0.0
django-cms==3.7.4
django-countries==6.1.3
django-elasticsearch-dsl==7.1.4
django-entangled==0.3
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-fsm==2.7.1
django-fsm-admin==1.2.4
django-ipware==3.0.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-parler==2.2
django-parler-rest==2.1
django-phonenumber-field==5.0.0
django-polymorphic==3.0.0
django-post-office==3.4.1
django-redis==4.12.1
django-redis-sessions==0.6.1
django-rest-auth==0.9.5
django-sass-processor==0.8.2
django-sekizai==2.0.0
django-select2==7.4.2
django-shop==1.2.3
django-treebeard==4.3.1
djangocms-admin-style==2.0.0
djangocms-bootstrap==1.1.2
djangocms-cascade==1.3.4
djangocms-text-ckeditor==4.0.0
djangorestframework==3.12.1
djangoshop-sendcloud==1.2.1
easy-thumbnails==2.7
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
geographiclib==1.50
geopy==2.0.0
html5lib==1.1
idna==2.10
isort==5.4.2
jsonfield==3.1.0
lazy-object-proxy==1.4.3
libsass==0.20.1
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.11
Pillow==8.0.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.11
webencodings==0.5.1
wrapt==1.12.1
